found matches due to swap
I figured out how to keep track of which tile matches were made due to a swap and which were just random chance. in the on_swap_function prior to updating the grid I just use get_to_free() to collect any matches that took place. this happens after a swap is made and before calling update_grid so any matches that exist in that in between space were defintiely due to the player