fixed rightmost column disappearing bug
for some reason I commented out a single line of code and then the bug got fixed. The bug was basically this: anything in the right most column would disappear when it hit the bottom. I guess someting was getting added to the to_free variable inside of get_to_free that wasn't supposed to be added idk honestly just commented out a line and now it works so idc haha